Split translator file, move translator initialization to main adapter file
